Reorganised ObjectBox backend into own directory
