+++ lib/src/backend/export_external.dart
@@ -2,8 +2,8 @@
 // A full license can be found at .\LICENSE
 
 export 'errors/errors.dart';
-export 'impls/web_noop/backend.dart'
-    if (dart.library.ffi) 'impls/objectbox/backend/backend.dart';
+export 'impls/objectbox/web_noop/backend.dart'
+    if (dart.library.ffi) 'impls/objectbox/native/backend/backend.dart';
 export 'interfaces/backend/backend.dart';
 export 'interfaces/backend/internal.dart';
 export 'interfaces/backend/internal_thread_safe.dart';
