+++ lib/src/backend/impls/objectbox/native/backend/internal_workers/thread_safe.dart
