+++ lib/src/backend/impls/objectbox/native/models/src/recovery_region.dart
@@ -6,7 +6,7 @@ import 'package:latlong2/latlong.dart';
 import 'package:meta/meta.dart';
 import 'package:objectbox/objectbox.dart';
 
-import '../../../../../../flutter_map_tile_caching.dart';
+import '../../../../../../../flutter_map_tile_caching.dart';
 
 /// Serialised [BaseRegion]
 @Entity()
